+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Change Log
 ==========
 
-## v0.2.0-dev
+## v0.2.0
 
 - fixed multi-page queries when `--num N` was greater that the initial result size
 - added `--np` and `--noprompt` options to just search and exit